ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

transformers==4.40.0

pdfminer.six==20221105

nltk==3.8.1

spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

numpy==1.26.4
pandas==2.2.1

scikit-learn==1.4.1.post1

python-multipart==0.0.9

pydantic==2.6.3
starlette==0.36.3
